pyproject.toml
[tool.poetry]
name = "comohay-alerts"
version = "0.0.1"
description = "Alerts notification system of ComoHay implemented with Typer, PyroGram and Supabase."
authors = ["Leynier Gutiérrez González <leynier41@gmail.com>"]
homepage = "https://github.com/leynier/comohay-alerts"
repository = "https://github.com/leynier/comohay-alerts"
readme = "README.md"

[tool.poetry.scripts]
comohay-alerts = "comohay_alerts.main:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
python-dotenv = "^0.19.0"
pydantic = "^1.8.2"
PyGithub = "^1.55"
beautifulsoup4 = "^4.9.3"
supabase-py = "^0.0.2"
requests = "2.25.1"
python-telegram-bot = "^13.7"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
